stu management readfilecsv

+++ C0323G1_le_huy_module2/src/additional_exercise/controller/StudentManagementController.java
@@ -21,22 +21,33 @@ public class StudentManagementController {
             System.out.println("3. Xem danh sách giảng viên hoặc học sinh");
             System.out.println("4. Thoát");
             System.out.println("Chọn chức năng: ");
+            do {
+                try {
+                    int option = Integer.parseInt(scanner.nextLine());
+                    if (option < 1 || option>4) {
+                        throw new RuntimeException("Bạn chỉ được nhập số từ 1 - 4. Vui lòng nhập lại");
+                    }
+                    switch (option) {
+                        case 1:
+                            studentManagementService.addStudentTeacher();
+                            break;
+                        case 2:
+                            studentManagementService.deleteStudentTeacher();
+                            break;
+                        case 3:
+                            studentManagementService.displayAll();
+                            break;
+                        case 4:
+                            System.out.println("Bạn đã nhấn thoát chương trình");
+                            System.exit(0);
+                    }
+                } catch (NumberFormatException numberFormatException) {
+                    System.out.println("Bạn chỉ được nhập số từ 1 - 4. Vui lòng nhập lại");
+                } catch (RuntimeException runtimeException) {
+                    System.out.println("Bạn chỉ được nhập số từ 1 - 4. Vui lòng nhập lại");
+                }
+            } while (true);
 
-            int option = Integer.parseInt(scanner.nextLine());
-            switch (option) {
-                case 1:
-                    studentManagementService.addStudentTeacher();
-                    break;
-                case 2:
-                    studentManagementService.deleteStudentTeacher();
-                    break;
-                case 3:
-                    studentManagementService.displayAll();
-                    break;
-                case 4:
-                    System.out.println("Bạn đã nhấn thoát chương trình");
-                    System.exit(0);
-            }
         }
         while (true);
     }
